requirements.txt
numpy
matplotlib
pandas
Pillow
pyxlsb
xlsxwriter
openpyxl
scipy